Csr signer ca configmap (#93)
Implemented a way of providing the CA certificates in a separate ConfigMap, with source of the data in the ConfigMap different for Openshift and k8s clusters.

Context:
Some Contrail components talk with each other using TLS. Their keys are generated and signed during pod's creation. That is done with CertficateSigningRequests k8s objects. The problem is in which key is used for signing these CSRs.

On a bare k8s cluster (f.e. one deployed using Kind), the defaults are used and everything is signed using the same CA (Certificate Authority) key. Also, all service account tokens hold value of this CA certificate.

On an Openshift cluster, a separate certificate chain is used for signing CSRs. It is stored in the csr-signer-ca ConfigMap in the openshift-kube-controller-manager-operator namespace.