Vmi management in contrail-provisioner (#201)
* Vmi management in contrail-provisioner

contrail-provisioner watches for changes in files mounted from ConfigMaps. In these files, different types of nodes that have to be registered in the contrail api server are listed.

To enable underlay->overlay (node->pod) connectivity, which is required for any readiness probe to pass, together with the VirtualRouter object, a VirtualMachineInterface object for the "vhost0" interface has to be created.

This PR adds this VMI creation/deletion functionality. I also tried to refactor the code and add a few unit tests.