Additional parameters for keystone
This change introduce new parameters for keystone resource:

  * auhtProtocol - 'http' or 'https' part of auth url which instructs
    clients to use encrypted connection or plain text
  * region - defines region used by keystone server and clients
  * projectDomain and userDomain - instructs clients which domain should
    be used

Those parameters are required when external keystone is in use to
properly configure clients in cluster. In that case keystone resource is
a source of information about external cluster and is not directly
managed by controler. It is external provisioner responsibility to mark
keyston resource active and set appropriate IP address.