Give possibility to mark keystone as external resource
With ExternalAddress defined for keystone CRD no other actions
will be taken by keyston controller than updating status of the
resource with information that it is external resource.
This information can be later used by keystone client to decide
how to connect to server, either using proxy or direct client.