Rollback after unsuccessful upgrade of Command (#435)
Rollback after unsuccessful upgrade of Command

This change fixes migration of Command's data during the upgrade, and also introduces rollback mechanism.
The idea behind this change is that during the migration a new DATABASE is created. This DATABASE is initialised with a new schema and then filled with migrated data. When the process is successful migrated DATABASE is renamed and put in the place of original one. The original one is saved as backup.

Co-authored-by: Katarzyna Bartkiewicz <kbartkiewicz@juniper.net>
Co-authored-by: Pawel Marchewka <pawel.marchewka@codilime.com>